Updated OrderController.cs
Added method async Task<ActionResult> PutOrderAsSubmitted(string), submits orders, before checks if there all products still available in stock before submiting. If some products is not available, then deletes it from order, recounting total price and informing the user about this and that the order needs to be double-checked and resubmit it.